==> 01-5_cleanBusiness.ipynb <==
import ast
import json
import random
import time

import autopep8
import numpy as np
import pandas as pd
from IPython.display import display
from pandas.api.types import is_numeric_dtype
from sklearn.preprocessing import LabelEncoder

# do get rid of annoying warnings
pd.set_option('future.no_silent_downcasting', True)

# global, for the print statements
verbose = True

# this should probably be on a seperate cell so I don't constantly reload the dataframes
fDefPath = 'reviews/yelpReviews/yelp_academic_dataset_'
# constants so I don't have to keep changing names
BS = 'business'
CH = 'checkin'
TI = 'tip'
RW = 'review'
US = 'user'

# subsets of what i care about
bssub = ['business_id', 'postal_code',
         'review_count', 'attributes', 'categories']
ussub = ['user_id', 'review_count', 'yelping_since']
rwsub = ['user_id', 'business_id', 'stars', 'text', 'date']

# constants for the file path
bspath = f'{fDefPath}{BS}.json'
chpath = f'{fDefPath}{CH}.json'
tipath = f'{fDefPath}{TI}.json'
rwpath = f'{fDefPath}{RW}.json'
uspath = f'{fDefPath}{US}.json'

# load the datasets
bsdf = pd.read_json(bspath, lines=True)
# take only what we want
bsdf = bsdf[bssub]

# create a manual dictionary for now
category_map = {
    # removed food, too broad, was catching grocery stores
    'restaurants': {'restaurants', 'sandwiches', 'american (new)', 'american (traditional)', 'pizza', 'bakeries', 'desserts', 'cafes', 'breakfast & brunch'},
    'retail': {'shopping', 'home & garden', 'fashion', 'real estate', 'grocery', 'department stores', 'electronics', 'toys'},
    'beauty': {'beauty & spas', 'hair salons', 'nail salons', 'massage', 'skin care', 'makeup artists'},
    'auto': {'automotive', 'auto repair'},
    'services': {'home services', 'local services', 'event planning & services'},
    'fitness': {'active life', 'fitness & instruction', 'gyms', 'yoga', 'martial arts'},
    'health': {'health & medical', 'doctors', 'dentists'},
    'pets': {'veterinarians', 'pet services', 'pets'},
    'nightlife': {'nightlife', 'bars', 'lounges', 'dance clubs', 'karaoke'},
    'travel': {'hotels & travel', 'hotels', 'tours'},
    'religious': {'religious organizations', 'synagogues'},
    'entertainment': {'festivals', 'arts & entertainment', 'television', 'mass media', 'museums', 'music venues', 'cinema', 'theaters'},
    'finance': {'banks & credit unions', 'financial services'},
    'legal': {'dui law', 'lawyers'},
    'government': {'public services & government', 'libraries'},
    'education': {'education'},
    'professional services': {'employment agencies', 'professional services'}
}

def matchBroadCategory(cat_list):
    for broad_cat, keywords in category_map.items():
        if any(cat in keywords for cat in cat_list):
            return broad_cat
    return "other"


# expand categories so we can then sort them into broader categories
bsdf = bsdf.dropna(subset=['categories', 'attributes']).copy()
bsdf['categories'] = bsdf['categories'].str.lower()
bsdf.loc[:, "categoryList"] = bsdf.loc[:, "categories"].str.split(", ")

# Flatten and count unique categories
allCategories = bsdf["categoryList"].explode().str.lower()
categoryCounts = allCategories.value_counts()

# Preview common categories
# print(categoryCounts.head(50))

bsdf["broadCategory"] = bsdf["categoryList"].apply(matchBroadCategory)

# put this into a csv so i can look at them better, only if verbose is True
if verbose:
    fileName = 'files/categorized.csv'
    bsdf[['categoryList', 'broadCategory']].to_csv(fileName)

# now I'm going to process the attributes in the business json
attDf = bsdf.dropna(subset=['attributes']).copy()

baseDf = bsdf[['business_id', 'attributes']].copy()
attDf = baseDf.attributes
# print(attDf.head())
attDf = pd.json_normalize(attDf, sep='_')

# normalize converts nested dictionaries into strings,
# convert it back into a dictionary
for col in attDf.columns:
    if attDf[col].apply(lambda x: isinstance(x, str) and x.strip().startswith('{')).any():
        attDf[col] = attDf[col].apply(
            lambda x: ast.literal_eval(x) if isinstance(x, str) else x)

# Find columns with nested dicts
nested_cols = [col for col in attDf.columns if isinstance(
    attDf.dropna(subset=[col])[col].iloc[0], dict)]

# Flatten each nested dict into new columns
for col in nested_cols:
    nested_df = pd.json_normalize(attDf[col])
    nested_df.columns = [f"{col}_{subcol}" for subcol in nested_df.columns]
    attDf = attDf.drop(columns=[col]).join(nested_df)

# merge attributes back with their business_id
# drop old attribute column
if len(baseDf.columns) > 1:
    baseDf = baseDf.drop(['attributes'], axis=1)

# remove index to make sure they line up
baseDf = baseDf.reset_index(drop=True)
attDf = attDf.reset_index(drop=True)
# merge
mergedDf = pd.concat([baseDf, attDf], axis=1)
if verbose:
    print(mergedDf.head())

# write at attributes normalized dataframe and the merged dataframe for
# easier manual checking (they're so long they don't fit in print statements)
if verbose:
    attDf.to_csv('files/att.csv')
    mergedDf.to_csv('files/mergedAtt.csv')

# doesn't manipulate data, just for checking. Will do nothing if verbose == False
if verbose:
    # check to make sure the business_id was lining up
    indices = list(range(len(baseDf)))
    random.shuffle(indices)
    # compare 15 random rows, using business_is as a common index
    lookat = indices[:15]

    # probably a cleaner way to do this, but oh well
    for ind in lookat:
        bsid = baseDf.iloc[ind]['business_id']
        mgid = mergedDf.iloc[ind]['business_id']
        bsAtt = list(bsdf.loc[bsdf['business_id'] == bsid]['attributes'])
        mdAtt = mergedDf.loc[mergedDf['business_id']
                             == mgid].drop(['business_id'], axis=1)
        mdAtt = mdAtt.dropna(axis=1)
        print(f' {bsid} : {mgid}')
        print(f'base frame att :\n {bsAtt} \n merged frame att: \n {mdAtt} ')

# see which columns are binary and which are categorical
# replace 'None' and None with np.nan
# replace true and false with binary
mergedDf.replace(['None', 'none', 'NONE', None], np.nan, inplace=True)
mergedDf.replace(['True', True, 'False', False], [1, 1, 0, 0], inplace=True)
# convert dtypes so is_numeric_dtype works (they're all object dtype right now)_
mergedDf = mergedDf.convert_dtypes()

for col in mergedDf.columns:
    if col == 'business_id':
        continue

    if verbose:
        print(
            f'{col} has attributes is of type {mergedDf[col].dtype}: {mergedDf[col].unique()} : trimming . . .')

    if is_numeric_dtype(mergedDf[col].dtype):
        if verbose:
            print('is numeric')
    else:
        mergedDf[col] = mergedDf[col].str.strip('u\'')

    if verbose:
        print(f'{col} has attributes: {mergedDf[col].unique()}\n\n')

# merged mergedDf (which is just business_id and  attributes(normalized) right now)
# with the broad categories
# drop unecessary columns (if they haven't been already)
if 'attributes' in bsdf.columns:
    bsdf = bsdf.drop(['attributes', 'categories', 'categoryList'], axis=1)

busCleanedDf = mergedDf.merge(bsdf, on='business_id', how='left')

label_encoder = LabelEncoder()
busCleanedDf['broadCategory'] = label_encoder.fit_transform(busCleanedDf['broadCategory'])

if verbose:
    pd.set_option('display.max_columns', None)
    display(busCleanedDf.head(3))
    pd.reset_option('display.max_columns')

# clean it up a bit, remove columns that are mostly empty
threshold = 0.9
busCleanedDf = busCleanedDf.loc[:, busCleanedDf.isnull().mean() < threshold]
busCleanedDf.to_csv('files/busCleaned.csv', index=False)
