==> 02-52_supportModelTraining.ipynb <==
import json
import string
import time
import gc
import glob
from io import StringIO

import joblib
import numpy as np
import pandas as pd
from IPython.display import display
from scipy.sparse import hstack
from sklearn.ensemble import RandomForestClassifier
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import classification_report
from sklearn.model_selection import train_test_split
from sklearn.svm import LinearSVC

pd.set_option('future.no_silent_downcasting', True)
verbose = True

zip_df_path = 'files/zip_cleaned_data_FULL.csv'
zip_df_chunks_path = 'files/review_chunks/zip_cleaned_data_0*.csv'

# Try stuff so I don't just keep getting errors if something goes wrong
try:
    print('Trying to load combined and cleaned YelpZip file')
    zip_df = pd.read_csv(zip_df_path)
    print('File found. File loaded into zip_df')
except FileNotFoundError:
    try:
        print('File not found. Trying to load cleaned file chunks.')
        chunk_files = sorted(glob.glob(zip_df_chunks_path))
        zip_df = pd.concat([pd.read_csv(file) for file in chunk_files], ignore_index = True)
        zip_df.to_csv('files/zip_cleaned_data_FULL.csv')
        print('Files combined and loaded into zip_df.')
    except Exception as e:
        print(f'Loading chunks failed.\nError: {e}')
except Exception as e:
    print(r'Something weird happened ¯\_(ツ)_/¯')
    print(f'The weird thing is: {e}')

print(zip_df.shape)
print(zip_df['label'].value_counts())

# structure of the dataframe should be [text, rating, label]
# vectorize with tf-idf
vizer = TfidfVectorizer()

x_text = vizer.fit_transform(zip_df['text'])

if verbose:
    print(f'Dataframe shape: {zip_df.shape}')
    print(f'Vectorized text shape: {x_text.shape}')

# combine the sparse matrix with the dense ratings column

# turn into 2d array
rate_feature = zip_df['rating'].values.reshape(-1, 1)

# combine vectorized text and ratings
# data
X = hstack([x_text, rate_feature])

# target labels
y = zip_df['label']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

c_params = [0.05, 0.1, 0.2, 0.5]
clf_names = []

# train the thing || trying differnt C score on SVM and Logistic Regression
models = {}

for c in c_params:
    # create a svm model with this c value
    # name = 'clf_svm' + str(c)
    # clf_names.append(name)
    # models[name] = LinearSVC(C=c)

    # t0 = time.time()
    # if verbose:
    #     print(f'Training {name}')
    # # actual training
    # models[name].fit(X_train, y_train)
    # if verbose:
    #     print(f'finished after {time.time() - t0} seconds')

    # create a logistic regression model with this c value
    name = 'clf_log' + str(c)
    clf_names.append(name)
    models[name] = LogisticRegression(C=c, max_iter=1000, class_weight='balanced')

    t0 = time.time()
    if verbose:
        print(f'Training {name}')
    # actual training
    models[name].fit(X_train, y_train)
    if verbose:
        print(f'finished after {time.time() - t0} seconds')


if verbose:
    print('Finished training')

for name in clf_names:
    y_pred = models[name].predict(X_test)
    print(f'Classification report for {name}')
    print(classification_report(y_test, y_pred))

# Best model for finding negative reviews is the logistic regression with c=0.1
# going to save that one and the vectorized (tf-idf)
# important because the model trained on this
# any new words might give me garbage.

# had tried svm's, they were bad at recalling the negative reviews.
# got higher accurary mainly by labeling everything real
joblib.dump(models['clf_log0.1'], 'models/support_log.pkl')
joblib.dump(vizer, 'models/support_vectorizer.pkl')
